Now its possible to save decimal numbers into CPU table and print them out

+++ dbo.CPU.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[CPU] (
     [Id]         INT IDENTITY (1, 1) NOT NULL,
-    [TotalClock] INT NULL,
+    [TotalClock] FLOAT NULL,
     PRIMARY KEY CLUSTERED ([Id] ASC)
 );
